requirements.txt
flask_login
flask_sqlalchemy
flask_wtf
flask_socketio
flask
gunicorn
psycopg2-binary
python-dotenv
gevent
eventlet
